+++ CHANGELOG.md
@@ -9,5 +9,6 @@
 * Events\Event
 * Exceptions\Exxx_EventStreamException
 * Exceptions\E4xx_EventStreamException
+* Exceptions\E4xx_NotAnEvent
 * Exceptions\E4xx_UnsupportedType
 * Streams\EventStream